Fix warnings

+++ TestManagerDConnection/DTXConnectionServices/DTXConnection.h
@@ -108,7 +108,7 @@ typedef void(^proxyHandler)(DTXProxyChannel *chan);
 
 // Remaining properties
 @property(readonly, copy) NSString *debugDescription;
-@property(readonly) unsigned long long hash;
+@property(readonly) NSUInteger hash;
 @property(readonly) Class superclass;
 
 @end
